[Prakash] - Refactored the code to add exception and java regex for validation of dates
