+++ lib/src/main/java/com/bridgelabz/hotelreservationsystem/main/FormatException.java
@@ -1,7 +1,10 @@
-package com.bridgelabz.hotelreservationsystem.test;
+package com.bridgelabz.hotelreservationsystem.main;
 
 public class FormatException extends Exception {
 	
+	public FormatException() {
+		
+	}
 	public FormatException(String message) {
 		System.out.println(message);
 	}
